Final/requirements.txt
transformers>=4.0.0
torch>=1.7.0
pandas
scikit-learn
tqdm
apify-client
python-dotenv==1.2.1
shap
numpy 
sentencepiece 
regex 
matplotlib
streamlit>=1.28.0